feat: Enable deep linking to public map view
Adds logic to redirect users to the public map view if an 'org' parameter is present in the URL. This facilitates direct access to organization-specific content via shared links.

Also, pre-populates the live feed in the public view with recent feedback items to improve initial user experience by preventing an empty state.